MC-18 remove load fragment
