+++ data/src/androidTest/java/com/malomnogo/data/RoomTest.kt
@@ -5,8 +5,8 @@ import androidx.room.Room
 import androidx.test.core.app.ApplicationProvider
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.malomnogo.data.core.CurrenciesDatabase
-import com.malomnogo.data.load.cache.CurrenciesDao
-import com.malomnogo.data.load.cache.CurrencyCache
+import com.malomnogo.data.currencies.cache.CurrenciesDao
+import com.malomnogo.data.currencies.cache.CurrencyCache
 import kotlinx.coroutines.runBlocking
 import org.junit.After
 import org.junit.Assert.assertEquals
