+++ data/src/main/java/com/malomnogo/data/core/CurrenciesDatabase.kt
@@ -2,10 +2,10 @@ package com.malomnogo.data.core
 
 import androidx.room.Database
 import androidx.room.RoomDatabase
+import com.malomnogo.data.currencies.cache.CurrenciesDao
+import com.malomnogo.data.currencies.cache.CurrencyCache
 import com.malomnogo.data.dashboard.cache.CurrencyPairCache
 import com.malomnogo.data.dashboard.cache.CurrencyPairDao
-import com.malomnogo.data.load.cache.CurrenciesDao
-import com.malomnogo.data.load.cache.CurrencyCache
 
 @Database(
     entities = [CurrencyCache::class, CurrencyPairCache::class],
