+++ data/src/main/java/com/malomnogo/data/currencies/cache/CurrenciesCacheDataSource.kt
@@ -1,4 +1,4 @@
-package com.malomnogo.data.load.cache
+package com.malomnogo.data.currencies.cache
 
 import javax.inject.Inject
