+++ data/src/main/java/com/malomnogo/data/currencies/cache/CurrencyCache.kt
@@ -1,4 +1,4 @@
-package com.malomnogo.data.load.cache
+package com.malomnogo.data.currencies.cache
 
 import androidx.room.ColumnInfo
 import androidx.room.Entity
