--- domain/src/main/java/com/malomnogo/domain/load/LoadCurrenciesRepository.kt
@@ -1,6 +1,0 @@
-package com.malomnogo.domain.load
-
-interface LoadCurrenciesRepository {
-
-    suspend fun fetchCurrencies(): LoadCurrenciesResult
-}
