--- domain/src/main/java/com/malomnogo/domain/load/LoadCurrenciesResult.kt
@@ -1,26 +1,0 @@
-package com.malomnogo.domain.load
-
-interface LoadCurrenciesResult {
-
-    fun map(mapper: Mapper)
-
-    interface Mapper {
-
-        fun mapSuccess()
-
-        fun mapError(message: String)
-    }
-
-    object Success : LoadCurrenciesResult {
-
-        override fun map(mapper: Mapper) {
-            mapper.mapSuccess()
-        }
-    }
-
-    data class Error(private val message: String) : LoadCurrenciesResult {
-        override fun map(mapper: Mapper) {
-            mapper.mapError(message)
-        }
-    }
-}
