--- presentation/src/main/java/com/malomnogo/presentation/load/LoadScreen.kt
@@ -1,5 +1,0 @@
-package com.malomnogo.presentation.load
-
-import com.malomnogo.presentation.main.Screen
-
-object LoadScreen : Screen.Replace(LoadFragment::class.java)
