--- presentation/src/main/res/layout/fragment_load.xml
@@ -1,15 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:id="@+id/loadRootLayout"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent">
-
-    <include
-        android:id="@+id/error"
-        layout="@layout/error" />
-
-    <include
-        android:id="@+id/progress"
-        layout="@layout/progress" />
-
-</FrameLayout>
